requirements.txt
numpy==1.20.1
tensorflow-cpu==2.1.0
scikit-image==0.18.1
Pillow==8.1.0
streamlit==0.79.0